build.config: Introduce FAST_BUILD for faster builds
If the environment variable FAST_BUILD is set, don't copy unstripped
kernel modules into the dist directory. For local kernel development,
this step is unnecessary because the unstripped kernel modules can still
be found under, for example, out/mixed/device-kernel/private/gs-google.
Copying these files to the dist directory only makes sense in a build
server environment where only files in the dist directory are retained.

Note that FAST_BUILD is independently interpreted by build.sh and
performs orthogonal optimizations to speed up the build for the purpose
of local kernel development. As of this writing, it decreases the lz4
compression ratio for faster compression and sets LTO=thin.

Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: I2ed3abab22df4caf30410a2c6e84f49b03ea406c